Rename tuft_function.h based on the function name instead of the return type.
